+++ tests/unit/test_main.py
@@ -326,7 +326,7 @@ import a
 import b
 """
     )
-    main.main([str(tmpdir), "--skip", "skip.py", "--check"])
+    main.main([str(tmpdir), "--extend-skip", "skip.py", "--check"])
 
     # without filter options passed in should successfully sort files
     main.main([str(python_file), str(should_skip), "--verbose", "--atomic"])