+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Find out more about isort's release policy [here](https://pycqa.github.io/isort/
   - Implemented #1669: Parallel (`-j`) now defaults to number of CPU cores if no value is provided.
   - Implemented #1668: Added a safeguard against accidental usage against /.
   - Implemented #1638 / #1644: Provide a flag `--overwrite-in-place` to ensure same file handle is used after sorting.
+  - Implemented #1684: Added support for extending skips with `--extend-skip` and `--extend-skip-glob`.
   - Documented #1685: Skip doesn't support plain directory names, but skip_glob does.
 
 ### 5.7.0 December 30th 2020